backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Rate Limiting & Caching
Flask-Limiter==3.5.0
redis==5.0.1

# Environment & Configuration
python-dotenv==1.0.0

# LLM Providers
openai==1.57.0
anthropic==0.40.0

# Security
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
cryptography==41.0.7

# Logging & Monitoring
python-json-logger==2.0.7

# Data Validation
pydantic==2.12.0
